updated task validation and added task exception

+++ src/main/java/myprojects/todolist/service/TaskService.java
@@ -3,6 +3,7 @@ package myprojects.todolist.service;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 import lombok.Setter;
+import myprojects.todolist.exception.TaskException;
 import myprojects.todolist.model.Task;
 import myprojects.todolist.repository.TasksRepository;
 import org.springframework.stereotype.Service;
@@ -18,7 +19,7 @@ import java.util.stream.Collectors;
 public class TaskService {
     private final TasksRepository tasksRepository;
 
-    public void saveTask(Task task) {
+    public void saveTask(Task task) throws TaskException {
         validateTask(task);
         tasksRepository.save(task);
     }
@@ -36,24 +37,22 @@ public class TaskService {
     }
 
     public void deleteTaskById(Long id) {
-        if (tasksRepository.findById(id).isPresent())
-            tasksRepository.deleteById(id);
+        if (tasksRepository.findById(id).isPresent()) tasksRepository.deleteById(id);
     }
 
-    public void validateTask(Task task) {
-        if (!task.getName().isBlank()) {
-            return;
-        } else if (task.getDescription().isBlank()) {
-            task.setName("Set main goal of your task :)");
-        } else {
+    public void validateTask(Task task) throws TaskException {
+        if (task.getName().isBlank() && task.getDescription().isBlank())
+            throw new TaskException("Task's title and description can't be blank", "Set title or description of your task");
+        else if (task.getName().isBlank()) {
             String[] words = task.getDescription().split(" ");
-            int splitPoint = Math.min(words.length / 2, 7);
-            task.setName(Arrays.stream(words, 0, splitPoint)
-                    .collect(Collectors.joining(" "))
-                    .trim());
-            task.setDescription(Arrays.stream(words, splitPoint, words.length)
-                    .collect(Collectors.joining(" "))
-                    .trim());
+            int descriptionLength = words.length;
+            int splitPoint = Math.min(descriptionLength / 2, 7);
+            task.setName(Arrays.stream(words, 0, splitPoint).collect(Collectors.joining("\\s+")).trim());
+            task.setDescription(Arrays.stream(words, splitPoint, descriptionLength).collect(Collectors.joining("\\s+")).trim());
         }
+        if (task.getName().length() > 255)
+            throw new TaskException("Title too long", "Title is too long. Max length is 255 charakters.");
+        if (task.getDescription().length() > 255)
+            throw new TaskException("Description too long", "Desription is too long. Max length is 255 charakters.");
     }
 }
